Separate sparse tensor representation from tucker.py
